Offer a way to disable tab-reloading at launch on Chromium-based browsers
Related discussion:
- https://github.com/gorhill/uBlock/commit/a0a9497b4aca86727b314d8fc31ad345dad08ac8#commitcomment-62560291

The behavior could be seen as annoying to some users, and before
this commit there is no way to opt-out of the behavior, introduced
in 1.40.0.

This commit is only intended for Chromium version of uBO, and
setting `suspendTabsUntilReady` to `no` will prevent active tab
reload at launch once filter lists are all loaded.

In next release, there will be a setting in the "Filter lists"
pane to easily control the behavior and `suspendTabsUntilReady`
will be removed.