Emphasize broken filtering at launch on toolbar icon badge
Related feedback:
- https://github.com/gorhill/uBlock/commit/769b8da664be2353bc2e3418789dc404bfde0a26#commitcomment-104695781

The incomplete filtering status of a given tab at browser launch
will be carried over visually as a yellowish `!` badge until the
web page in the tab is force reloaded, navigated away, or closed.

The purpose is to make it obvious to end users that a web page
has not been filtered properly and to avoid issue reports
related to this.

It is expected that Firefox should never be affected by cases of
yellowish badge -- that is unless the setting "Suspend network
activity [...]" has been disabled, in which case the new behavior
will also be useful to those who disabled the setting.